selenide configuration moved to BaseTest
