+++ src/test/java/iteration2/ui/BaseTest.java
@@ -1,0 +1,20 @@
+package iteration2.ui;
+
+import com.codeborne.selenide.Configuration;
+import org.junit.jupiter.api.BeforeAll;
+
+import java.util.Map;
+
+public class BaseTest {
+    @BeforeAll
+    public static void setupSelenoid() {
+        Configuration.remote = "http://localhost:4444/wd/hub";
+        Configuration.baseUrl = "http://192.168.70.184:3000";
+        Configuration.browser = "chrome";
+        Configuration.browserSize = "1920x1080";
+
+        Configuration.browserCapabilities.setCapability(
+                "selenoid:options",
+                Map.of("enableVNC", true, "enableLog", true));
+    }
+}
