+++ src/test/java/iteration2/ui/DepositMoneyTest.java
@@ -1,20 +1,39 @@
 package iteration2.ui;
 
-import com.codeborne.selenide.Configuration;
+import com.codeborne.selenide.Selenide;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import steps.User;
 
-import java.util.Map;
+import static com.codeborne.selenide.Selenide.executeJavaScript;
 
-public class DepositMoneyTest {
+public class DepositMoneyTest extends BaseTest {
+    private static User user;
+
     @BeforeAll
-    public static void setupSelenoid() {
-        Configuration.remote = "http://localhost:4444/wd/hub";
-        Configuration.baseUrl = "http://192.168.0.216:3000";
-        Configuration.browser = "chrome";
-        Configuration.browserSize = "1920x1080";
+    public static void createUsersAndAccounts() {
+        // Создание пользователя и счета
+        user = new User.Builder()
+                .createRandomUser()
+                .createFirstAccount()
+                .build();
+//        new ValidationRequester(Endpoint.LOGIN, RequestSpecs.authAsAdmin(), ResponseSpecs.returnsOk())
+//                .send(new LoginRequest("admin", "admin"));
+    }
 
-        Configuration.browserCapabilities.setCapability(
-                "selenoid:options",
-                Map.of("enableVNC", true, "enableLog", true));
+    // Удаляем юзера после прохождения всех тестов
+    @AfterAll
+    public static void deleteUsers() {
+        user.deleteUser();
+    }
+
+    @Test
+    @DisplayName("Успешное пополнение баланса")
+    public void userCanDepositTest() {
+        executeJavaScript("localStorage.setItem('authToken', arguments[0]);", user.token());
+
+        Selenide.open("/dashboard");
     }
 }
